HistoryFlowPanel.form ok

+++ JFlowMonitor/src/Database/Flow.java
@@ -14,8 +14,8 @@ import java.util.Date;
  */
 public class Flow {
     public Date sDate;
-    public int innerSize;
-    public int outerSize;
+    public long innerSize;
+    public long outerSize;
     public Flow(){
         innerSize = 0;
         outerSize = 0;
